refactor: useNotification & useProducts 훅 추가
