+++ src/basic/hooks/useNotification.ts
@@ -1,0 +1,24 @@
+import { useState, useCallback } from 'react';
+import { Notification } from '../../types';
+
+export function useNotification() {
+  const [notifications, setNotifications] = useState<Notification[]>([]);
+
+  const addNotification = useCallback(
+    (message: string, type: 'error' | 'success' | 'warning' = 'success') => {
+      const id = Date.now().toString();
+      setNotifications((prev) => [...prev, { id, message, type }]);
+
+      setTimeout(() => {
+        setNotifications((prev) => prev.filter((n) => n.id !== id));
+      }, 3000);
+    },
+    [],
+  );
+
+  return {
+    notifications,
+    setNotifications,
+    addNotification,
+  };
+}
